// apps/web/src/app/register/enum.ts
export enum School {
    BSSM = "BSSM",
    DGSM = "DGSM",
    DSM = "DSM",
    GSM = "GSM",
}

export enum Major {
    SOFTWARE = "SOFTWARE",
    EMBEDDED = "EMBEDDED",
    SECURITY = "SECURITY",
    IOT = "IOT",
    AI = "AI",
}

export enum EducationStatus {
    ENROLLED = "ENROLLED",
    GRADUATED = "GRADUATED",
}